# Add Webpack, Babel, and Modern ES6 Syntax Support