--- a/netlify.toml
+++ b/netlify.toml
@@ -1,0 +1,6 @@
+[build]
+  command = "npm run build"
+  publish = "public"
+
+[context.deploy-preview]
+  command = "npm run build:preview"