--- a/site/config.toml
+++ b/site/config.toml
@@ -1,4 +1,4 @@
-baseURL = "https://tour.dgraph.io"
+baseURL = ""
 languageCode = "en-us"
 title = "A Tour of Dgraph"
 theme="hugo-tutorials"