--- a/site/themes/hugo-tutorials/layouts/partials/footer.html
+++ b/site/themes/hugo-tutorials/layouts/partials/footer.html
@@ -1,0 +1,13 @@
+    </main>
+
+    {{ $script := .Site.Data.webpack.main }}
+    {{ with $script.js }}
+    <script src="{{ relURL . }}"></script>
+    {{ end }}
+
+    <script>
+      hljs.initHighlightingOnLoad();
+    </script>
+    </body>
+
+    </html>