--- a/src/index.js
+++ b/src/index.js
@@ -1,0 +1,11 @@
+// CSS
+import './css/main.css';
+
+// JavaScript
+import './js/clipboard.min.js';
+import './js/dgraph.js';
+import './js/runnable.js';
+import './js/runnable-url/index.js';
+
+// Say hello
+console.log('🦊 Hello! Edit me in src/index.js');