--- a/webpack.dev.js
+++ b/webpack.dev.js
@@ -1,0 +1,42 @@
+const merge = require('webpack-merge');
+const path = require('path');
+const CleanWebpackPlugin = require('clean-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+const common = require('./webpack.common');
+
+module.exports = merge(common, {
+  mode: 'development',
+
+  output: {
+    filename: '[name].js',
+    chunkFilename: '[id].css'
+  },
+
+  devServer: {
+    port: process.env.PORT || 3000,
+    contentBase: path.join(process.cwd(), './public'),
+    watchContentBase: true,
+    stats: 'none',
+    quiet: false,
+    open: true,
+    historyApiFallback: {
+      rewrites: [{ from: /./, to: '404.html' }]
+    }
+  },
+
+  plugins: [
+    new CleanWebpackPlugin({
+      cleanOnceBeforeBuildPatterns: [
+        'public/**/*.js',
+        'public/**/*.css',
+        'site/content/webpack.json'
+      ]
+    }),
+
+    new MiniCssExtractPlugin({
+      filename: '[name].css',
+      chunkFilename: '[id].css'
+    })
+  ]
+});