@MichelDiz 2019-05-30
https://www.browserling.com/browse/win/7/ie/11/https%3A%2F%2Fdetermined-stallman-63e0ba.netlify.com%2F

I've test it in some old browsers using browserling and is working great.

BTW, to test locally you need to:
```
npm install
npm run build
npm run start
```
or
```
yarn install
yarn build
yarn start
```