@MichelDiz 2019-09-10
I'm closing this due to the new directions the Tour is taking.

Thanks for the effort!